Authorization

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require("express");
 
 const Post = require('../models/post');
 const multer = require('multer');
+const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
 
@@ -27,12 +28,13 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post("", multer({storage: storage}).single("image") ,(req,res,next)=>{
+router.post("", checkAuth ,multer({storage: storage}).single("image") ,(req,res,next)=>{
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(CreatedPost => {
     res.status(201).json({
@@ -66,7 +68,7 @@ router.get('',(req, res, next)=>{
   });
 });
 
-router.put("/:id",  multer({storage: storage}).single("image") ,(req, res, next)=>{
+router.put("/:id", checkAuth, multer({storage: storage}).single("image") ,(req, res, next)=>{
   let imagePath = req.body.imagePath;
   if(req.file){
     const url = req.protocol + '://' + req.get("host");
@@ -76,13 +78,20 @@ router.put("/:id",  multer({storage: storage}).single("image") ,(req, res, next)
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id},post).then(result => {
-    res.status(200).json({
-      message:'Edit successfull!'
-    });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId },post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message:'Edit successfull!'
+      });
+    } else {
+      res.status(401).json({
+        message:'Not Authorized!'
+      });
+    }
+
   });
 });
 
@@ -98,12 +107,18 @@ router.get("/:id",( req, res, next ) =>{
   });
 });
 
-router.delete("/:id",(req, res, next)=>{
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Post Deleted'
+router.delete("/:id", checkAuth,(req, res, next)=>{
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message:'Deletion successfull!'
       });
+    } else {
+      res.status(401).json({
+        message:'Not Authorized!'
+      });
+    }
+
   });
 });
